Move LogsScreen to /support
